+++ src/components/Main.js
@@ -1,9 +1,8 @@
 import { BrowserRouter } from "react-router-dom";
+import FetchController from "./FetchController";
 import NavBarController from "./nav/NavBarController";
 import NavContent from "./nav/NavContent";
 import './main.css'
-import FetchController from "./FetchController";
-import Fetch from "./Fetch";
 
 export default function Main() {
     return (
@@ -13,7 +12,7 @@ export default function Main() {
                 <div className="nav-content">
                     <NavContent />
                 </div>
-                <FetchController><Fetch/></FetchController>
+                <FetchController/>
             </div>
         </BrowserRouter>
     )
