+++ src/components/assignments/Assignments.js
@@ -1,10 +1,11 @@
 import { useContext, useEffect } from "react"
+import { ActionType } from "../../state/State";
 import { StateContext } from "../../state/StateContext.tsx"
 
 export default function Assignments({ location }) {
     const { state, actions } = useContext(StateContext);
     useEffect(() => {
-        actions.fetch("assignments", [state.provider.name], "assignments")
+        actions.fetch("assignments", ActionType.ASSIGNMENTS, [state.provider.name]);
     }, [location.key, state.provider.name])
     return (<div>
         <h4>Assignments</h4>
