+++ src/components/commits/Commits.js
@@ -1,10 +1,11 @@
 import { useContext, useEffect } from "react"
+import { ActionType } from "../../state/State";
 import { StateContext } from "../../state/StateContext.tsx"
 
 export default function Commits({ location }) {
     const { state, actions } = useContext(StateContext);
     useEffect(() => {
-        actions.fetch("commits", [state.provider.name], "commits")
+        actions.fetch("commits", ActionType.COMMITS, [state.provider.name])
     }, [location.key, state.provider.name])
     return (<div>
         <h4>Commits</h4>
