+++ src/components/startup/StartUp.tsx
@@ -1,7 +1,6 @@
 import { useContext } from "react";
-import { StateContext } from "../../state/StateContext.tsx";
+import { StateContext } from "../../state/StateContext";
 import Main from "../Main";
-import FetchController from "../FetchController";
 import SplashController from "./SplashController";
 
 export default function StartUp() {
@@ -9,7 +8,7 @@ export default function StartUp() {
     return (
         <>
             {
-                state.ready ? <Main /> : <FetchController><SplashController/></FetchController>
+                state.ready ? <Main /> : <SplashController/>
             }
         </>
     );
